// web/src/kiosk/pages/__generated__/StatusQuery.graphql.ts
export type StatusQuery$variables = {
  first: number;
};
export type StatusQuery$data = {
  readonly session: {
    readonly location: {
      readonly periods: {
        readonly edges: ReadonlyArray<{
          readonly node: {
            readonly id: string;
            readonly person: {
              readonly firstName: string;
              readonly id: string;
              readonly lastName: string;
            };
            readonly startTime: number;
          };
        }>;
      };
    };
  };
};
export type StatusQuery = {
  response: StatusQuery$data;
  variables: StatusQuery$variables;
};